refactor: Update colors for display_path and display_user functions
This commit updates the colors used in the display_path and display_user functions in the gears_utils.c file. The new colors improve the visual appearance of the displayed information. This change enhances the user interface of the minishell by providing a more visually appealing display of the current path and user information.

+

solved some var bugs