refactor: Add echon, pwd, and unset functions
This commit adds the echon, pwd, and unset functions to the minishell. The echon function is used to print a string without a newline character, the pwd function is used to print the current working directory, and the unset function is used to remove an environment variable. These functions enhance the functionality of the minishell by providing additional capabilities for printing and manipulating environment variables.